// lab7/frontend/src/api/Api.ts
export interface SerializerBigORequestJSON {
  bigo_request_id?: number;
  calculated_complexity?: string;
  calculated_time?: number;
  creator_login?: string;
  date_create?: string;
  date_finish?: string;
  date_update?: string;
  moderator_login?: string;
  status?: string;
}
// ...
export interface AllBigoRequestsListParams {
  /** Начальная дата (YYYY-MM-DD) */
  "from-date"?: string;
  /** Конечная дата (YYYY-MM-DD) */
  "to-date"?: string;
  /** Статус заявки */
  status?: string;
}